requirements.txt
aiohttp==3.9.3
orjson==3.9.13
mysql-connector-python==8.3.0
pytest==8.0.0
pytest-asyncio 0.23.5